* release/1.1.0:
  ansible-tomcat 1.1.0: Various improvements
  Add vagrant environments for CentOS 6 and 7
  Allow to add environment files for vagrant based development
  [ansible-generator] Do not store .retry files when testing this role
  Fix user to run tomcat as with SysV default service template
  use JSON format for facts.j2 template and expose some instances configurations inside facts
  Update README.md to add new option: tomcat_facts_template
  add missing catalina_base directories: bin, lib and work

+++ rake/vagrant.rb
@@ -4,7 +4,11 @@ require 'rspec/core/rake_task'
 
 desc "Bring up Vagrant VM"
 task :up do
-  sh %{vagrant up --no-provision}
+  if ENV['ANSIBLE_TOMCAT_VAGRANT_PROVIDER']
+    sh 'vagrant', 'up', '--no-provision', '--provider', ENV['ANSIBLE_TOMCAT_VAGRANT_PROVIDER']
+  else
+    sh %{vagrant up --no-provision}
+  end
 end
 
 
@@ -16,8 +20,10 @@ end
 
 desc "Cleanup Vagrant VM environment"
 task :clean do
-  sh %{vagrant halt}
-  sh %{vagrant destroy --force}
+  if not ENV['RAKE_ANSIBLE_VAGRANT_DONT_CLEANUP'] == '1'
+    sh %{vagrant halt}
+    sh %{vagrant destroy --force}
+  end
 end
